// pay-bridge-web-app/src/Interfaces/bankAssetModel.ts
export default interface bankAsset {
  assetId: number;
  assetUniqueNumber: string;
  ibaN_Number: string;
  currencyType: string;
  balance: number;
  isActive: boolean;
  status: string;
  bankName: string;
  bankAccountUniqueNumber: string;
  shortCompanyOwnerName: string;
  registrationDate: string;
}